LayoutTests/media/media-webm-no-duration-expected.txt


RUN(audio.src = "../../../media/content/test-no-duration.webm")
EVENT(canplay)
RUN(promise = audio.play())
Promise resolved OK
EXPECTED (audio.currentTime > '0') OK
END OF TEST
